[add/fix] エラーの出力およびdocコメントの追加

+++ models/Category.php
@@ -3,23 +3,38 @@
 namespace app\models;
 
 use app\core\Model;
-use app\models\Memo;
 
 
 class Category extends Model
 {
+  /**
+   * カテゴリーの新規登録
+   *
+   * @param string $user_id
+   * @param array $data
+   * @return boolean
+   */
+
   public function insert($user_id,$data)
   {
     $sql = "INSERT INTO categories (user_id,name) VALUES(:user_id,:cat_name)";
 
     $this->prepare($sql);
-    $this->bind(':user_id',$user_id);
+    $this->bind(':user_id',$user_id,\PDO::PARAM_INT);
     $this->bind(':cat_name',$data['cat_name']);
 
     return $this->execute();
   }
 
-  public function getCategories($user_id)
+
+  /**
+   * ログインユーザの全カテゴリーを取得
+   *
+   * @param string $user_id
+   * @return array
+   */
+
+  public function getAllCategories($user_id)
   {
     $sql = "SELECT * FROM categories WHERE user_id = {$user_id}";
 
@@ -34,49 +49,81 @@ class Category extends Model
     }
   }
 
-  public function findOne($user_id,$id)
+
+  /**
+   * ログインユーザが持つ特定のカテゴリーの取得
+   * 
+   * @param string $user_id
+   * @param string $id
+   * @return array
+   */
+
+
+  public function getCategory($user_id,$id)
   {
     $sql = "SELECT * FROM categories WHERE user_id = {$user_id} AND id = :id";
 
     
 
     $this->prepare($sql);
-    $this->bind(":id",$id);
+    $this->bind(":id",$id,\PDO::PARAM_INT);
 
     $this->execute();
 
     $data = $this->fetch(\PDO::FETCH_ASSOC);
 
     if($data === false){
-      die('不正なリクエスト');
+      $this->handler->output("不正なリクエスト");
     }
     
     return $data;
   }
 
+
+  /**
+   * ログインユーザのカテゴリの更新
+   *
+   * @param string $user_id
+   * @param string $id
+   * @param array $data
+   * @return boolean
+   */
+
   public function update($user_id,$id,$data)
   {
-    $sql = "UPDATE categories SET name = :name WHERE user_id = {$user_id} AND id = :id";
-
-    $this->prepare($sql);
-    $this->bind(":name",$data['cat_name']);
-    $this->bind(":id",$id);
-
-    return $this->execute();
+    //該当レコードが存在するか
+    if($this->getCategory($user_id,$id)){
+      $sql = "UPDATE categories SET name = :name,updated_at = :updated_at WHERE user_id = {$user_id} AND id = :id";
+  
+      $this->prepare($sql);
+      $this->bind(":name",$data['cat_name']);
+      $this->bind(":updated_at",$this->today);
+      $this->bind(":id",$id,\PDO::PARAM_INT);
+  
+      return $this->execute();
+    }
   }
 
+
+  /**
+   * ログインユーザが持つカテゴリの削除
+   *
+   * @param string $user_id
+   * @param string $id
+   * @return boolean
+   */
+  
   public function delete($user_id,$id)
   {
-    $sql = "DELETE FROM categories WHERE user_id = {$user_id} AND id = :id";
+    //該当レコードが存在するか
+    if($this->getCategory($user_id,$id)){
+      $sql = "DELETE FROM categories WHERE user_id = {$user_id} AND id = :id";
 
-    $this->prepare($sql);
-    $this->bind(":id",$id);
+      $this->prepare($sql);
+      $this->bind(":id",$id,\PDO::PARAM_INT);
 
-    if($this->execute()){
-      return true;
+      return $this->execute();
     }
-
-    return false;
   }
 
 }
